use double instead of float for deltaTime, perform velocity interpolation, prevent CollisionSystem from looping on already colliding entities, assign priority to EntitySystems, calculate offset and clearRect() before rendering, use primitive integer instead of Integer for PrioritizedPoint2D.priority, rename Renderable.draw() to render()

+++ src/core/EntitySystem.java
@@ -39,5 +39,5 @@ public abstract class EntitySystem implements Comparable<EntitySystem> {
 		return 1;
 	}
 
-	public abstract void update(float deltaTime);
+	public abstract void update(double deltaTime);
 }
